address reviewer comments
